pyproject.toml
[project]
name = "chatbot_api"
version = "0.1"
requires-python = ">=3.8" 
dependencies = [
    "asyncio==3.4.3",
    "fastapi==0.109.0",
    "langchain==0.2.10",
    "langchain-openai==0.1.17",
    "langchain-community==0.2.9",
    "langchainhub==0.1.14",
    "neo4j==5.14.1",
    "numpy==1.26.2",
    "openai==1.36.0",
    "opentelemetry-api==1.22.0",
    "pydantic==2.5.1",
    "uvicorn==0.25.0",
    "six==1.16.0",
    "python-dotenv"             # Explicitly add six

]

[project.optional-dependencies]
dev = ["black", "flake8"]
